Split integration with DynamoDb by files
