+++ products-service/src/integrations/dynamo-db/connection.ts
@@ -1,0 +1,7 @@
+import * as AWS from "aws-sdk";
+
+AWS.config.credentials = new AWS.SharedIniFileCredentials({
+  profile: "default",
+});
+
+export const dynamoDbClient = new AWS.DynamoDB.DocumentClient();
