+++ products-service/src/integrations/dynamo-db/index.ts
@@ -1,0 +1,4 @@
+export * from "./connection";
+export * from "./functions/get-all-products";
+export * from "./functions/get-product-by-id";
+export * from "./functions/insert-product";
